chore: move `INIT_LESSON` handler
from project.py to course.py
